+++ backend/src/PetProject.Domain/Entities/PetPhoto.cs
@@ -1,9 +1,10 @@
-﻿namespace PetProject.Domain
+﻿namespace PetProject.Domain.Entities
 {
     public class PetPhoto
     {
         public Guid Id { get; private set; }
-        public string Path { get; private set; }
+
+        public string Path { get; private set; } = default!;
         public bool IsMainPhoto { get; private set; }
     }
 }
