+++ backend/src/PetProject.Domain/Entities/SotialNetwork.cs
@@ -1,4 +1,4 @@
-﻿namespace PetProject.Domain
+﻿namespace PetProject.Domain.Entities
 {
     public struct SotialNetwork
     {
