+++ backend/src/PetProject.Domain/ValueObjects/HelpStatus.cs
@@ -1,0 +1,6 @@
+﻿public enum HelpStatus
+{
+    NeedHelp,
+    LookingForAHome,
+    FindAHome
+}
